Put together the first stable version of the app
Implemented a responsive Grid to display the collection of books, as
well as many Flexbox features in order to react more smoothly
to viewport changes as well as difference device sizes.

Tinkered with modals for a while until managed to pull out a
nice little "new book" pop-up, which darkens the background and can be
exited by clicking outside of the box.

Customized form controls in general, especially the "read" checkbox,
which was stylized as a toggle slider.

Make use of an object constructor to represent each book and work
around that pattern.